Migrate APIs out of StorageManager: load_and_delete_update_conditions.
