+++ tiledb/sm/storage_manager/storage_manager_canonical.h
@@ -77,9 +77,7 @@ class GroupDetails;
 class Metadata;
 class MemoryTracker;
 class Query;
-class QueryCondition;
 class RestClient;
-class UpdateValue;
 class VFS;
 
 enum class EncryptionType : uint8_t;
@@ -255,18 +253,6 @@ class StorageManagerCanonical {
    * @return Status
    */
   Status is_group(const URI& uri, bool* is_group) const;
-
-  /**
-   * Loads the delete and update conditions from storage.
-   *
-   * @param opened_array The opened array.
-   * @return Status, vector of the conditions, vector of the update values.
-   */
-  tuple<
-      Status,
-      optional<std::vector<QueryCondition>>,
-      optional<std::vector<std::vector<UpdateValue>>>>
-  load_delete_and_update_conditions(const OpenedArray& opened_array);
 
   /** Removes a TileDB object (group, array). */
   Status object_remove(const char* path) const;
